move copilot SDK files to pipelines/copilot_agents/, remove slop prompt
- move copilot_agent.py -> pipelines/copilot_agents/
- move copilot_event_adapter.py -> pipelines/copilot_agents/
- move tools/ -> pipelines/copilot_agents/tools/
- remove _NO_FAKE_TOOL_USE_INSTRUCTION (allowlist handles this)
- update all imports across src and tests
