+++ src/archi/pipelines/copilot_agents/__init__.py
@@ -1,0 +1,5 @@
+"""Copilot SDK agent package."""
+
+from .copilot_agent import CopilotAgentPipeline
+
+__all__ = ["CopilotAgentPipeline"]
