+++ src/archi/pipelines/copilot_agents/copilot_event_adapter.py
