+++ src/archi/pipelines/copilot_agents/tools/document_collector.py
